Moved crypto functions into their own class
